#64 fix: click on the avatar to move to the profile page
