+++ frontend/src/App.tsx
@@ -1,268 +1,59 @@
+import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { ChakraProvider } from "@chakra-ui/react";
+import {
+  ChakraProvider,
+  Box,
+  useColorModeValue,
+  Drawer,
+  DrawerContent,
+} from "@chakra-ui/react";
 import Home from "./routes/Home";
 import AccessHistory from "./routes/AccessHistory";
 import Profile from "./routes/Profile";
 import Footer from "./features/Footer";
-
-import {
-  IconButton,
-  Avatar,
-  Box,
-  CloseButton,
-  Flex,
-  HStack,
-  VStack,
-  Icon,
-  useColorModeValue,
-  Text,
-  Drawer,
-  DrawerContent,
-  useDisclosure,
-  BoxProps,
-  FlexProps,
-  Menu,
-  MenuButton,
-  MenuDivider,
-  MenuItem,
-  MenuList,
-  Heading,
-} from "@chakra-ui/react";
-import {
-  FiHome,
-  FiTrendingUp,
-  FiCompass,
-  FiStar,
-  FiSettings,
-  FiMenu,
-  FiBell,
-  FiChevronDown,
-  FiBarChart,
-  FiBarChart2,
-  FiMapPin,
-} from "react-icons/fi";
-import { FaHistory } from "react-icons/fa";
-import { GiPerspectiveDiceSixFacesRandom } from "react-icons/gi";
-import { IconType } from "react-icons";
-
-interface LinkItemProps {
-  name: string;
-  icon: IconType;
-  href: string;
-}
-
-interface NavItemProps extends FlexProps {
-  icon: IconType;
-  href: string;
-  children: React.ReactNode;
-}
-
-interface MobileProps extends FlexProps {
-  onOpen: () => void;
-}
-
-interface SidebarProps extends BoxProps {
-  onClose: () => void;
-}
-
-const LinkItems: Array<LinkItemProps> = [
-  { name: "Attendee List", icon: FiHome, href: "/" },
-  { name: "Access History", icon: FaHistory, href: "/access-history" },
-  { name: "Ranking", icon: FiBarChart2, href: "/" },
-  { name: "Gacha", icon: GiPerspectiveDiceSixFacesRandom, href: "/" },
-  { name: "ISDL Map", icon: FiMapPin, href: "/" },
-  { name: "Settings", icon: FiSettings, href: "/" },
-];
-
-const SidebarContent = ({ onClose, ...rest }: SidebarProps) => {
-  return (
-    <Box
-      transition="3s ease"
-      bg={useColorModeValue("white", "gray.900")}
-      borderRight="1px"
-      borderRightColor={useColorModeValue("gray.200", "gray.700")}
-      w={{ base: "full", md: 60 }}
-      pos="fixed"
-      h="full"
-      zIndex="0" // SidebarContentを前面に表示
-      {...rest}
-    >
-      <Flex h="6" alignItems="center" mx="8" justifyContent="space-between">
-        <CloseButton display={{ base: "flex", md: "none" }} onClick={onClose} />
-      </Flex>
-      {LinkItems.map((link) => (
-        <NavItem key={link.name} icon={link.icon} href={link.href}>
-          {link.name}
-        </NavItem>
-      ))}
-    </Box>
-  );
-};
-
-const NavItem = ({ icon, href, children, ...rest }: NavItemProps) => {
-  return (
-    <Box
-      as="a"
-      href={href}
-      style={{ textDecoration: "none" }}
-      _focus={{ boxShadow: "none" }}
-    >
-      <Flex
-        align="center"
-        p="4"
-        mx="4"
-        borderRadius="lg"
-        role="group"
-        cursor="pointer"
-        _hover={{
-          bg: "cyan.400",
-          color: "white",
-        }}
-        {...rest}
-      >
-        {icon && (
-          <Icon
-            mr="4"
-            fontSize="20"
-            _groupHover={{
-              color: "white",
-            }}
-            as={icon}
-          />
-        )}
-        {children}
-      </Flex>
-    </Box>
-  );
-};
-
-const MobileNav = ({ onOpen, ...rest }: MobileProps) => {
-  return (
-    <Flex
-      position="fixed"
-      top={0}
-      left={0}
-      right={0}
-      zIndex="2" // SidebarContentを前面に表示
-      px={{ base: 4, md: 4 }}
-      height="20"
-      alignItems="center"
-      bg={useColorModeValue("rgba(79, 209, 197, 1)", "gray.900")}
-      borderBottomColor={useColorModeValue("gray.200", "gray.700")}
-      justifyContent={{ base: "space-between", md: "space-between" }}
-      {...rest}
-    >
-      <IconButton
-        display={{ base: "flex", md: "none" }}
-        onClick={onOpen}
-        variant="outline"
-        aria-label="open menu"
-        icon={<FiMenu />}
-      />
-
-      <Heading
-        as="h1"
-        fontSize={{ base: "xl", md: "4xl" }}
-        ml={4}
-        color="white"
-      >
-        ISDL Sentinel
-      </Heading>
-
-      <HStack spacing={{ base: "0", md: "6" }}>
-        <IconButton
-          size="lg"
-          variant="ghost"
-          aria-label="open menu"
-          icon={<FiBell />}
-        />
-        <Flex alignItems={"center"}>
-          <Menu>
-            <MenuButton
-              py={2}
-              transition="all 0.3s"
-              _focus={{ boxShadow: "none" }}
-            >
-              <HStack>
-                <Avatar
-                  size={"sm"}
-                  src={
-                    "https://images.unsplash.com/photo-1619946794135-5bc917a27793?ixlib=rb-0.3.5&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&s=b616b2c5b373a80ffc9636ba24f7a4a9"
-                  }
-                />
-                <VStack
-                  display={{ base: "none", md: "flex" }}
-                  alignItems="flex-start"
-                  spacing="1px"
-                  ml="2"
-                >
-                  <Text fontSize="sm">Hayato Oka</Text>
-                  <Text fontSize="xs" color="gray.600">
-                    Admin
-                  </Text>
-                </VStack>
-                <Box display={{ base: "none", md: "flex" }}>
-                  <FiChevronDown />
-                </Box>
-              </HStack>
-            </MenuButton>
-            <MenuList
-              bg={useColorModeValue("white", "gray.900")}
-              borderColor={useColorModeValue("gray.200", "gray.700")}
-            >
-              <MenuItem as="a" href="/profile">
-                Profile
-              </MenuItem>
-              <MenuItem>Change Password</MenuItem>
-              <MenuDivider />
-              <MenuItem>Sign out</MenuItem>
-            </MenuList>
-          </Menu>
-        </Flex>
-      </HStack>
-    </Flex>
-  );
-};
+import SidebarContent from "./features/SidebarContent";
+import MobileNav from "./features/MobileNav";
+import { useDisclosure } from "@chakra-ui/react";
 
 const App = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
+
   return (
     <ChakraProvider>
-      <Box minH="100vh" bg={useColorModeValue("gray.100", "gray.900")}>
-        {/* mobilenav */}
-        <MobileNav onOpen={onOpen} />
-        <SidebarContent
-          onClose={() => onClose}
-          display={{ base: "none", md: "block" }}
-          pt={20}
-        />
-        <Drawer
-          isOpen={isOpen}
-          placement="left"
-          onClose={onClose}
-          returnFocusOnClose={false}
-          onOverlayClick={onClose}
-          size="full"
-        >
-          <DrawerContent>
-            <SidebarContent onClose={onClose} />
-          </DrawerContent>
-        </Drawer>
-        <Box
-          pl={{ base: 2, md: 64 }}
-          pr={{ base: 2, md: 6 }}
-          pt={{ base: 2, md: 24 }}
-        >
-          <BrowserRouter>
+      <BrowserRouter>
+        <Box minH="100vh" bg={useColorModeValue("gray.100", "gray.900")}>
+          <MobileNav onOpen={onOpen} />
+          <SidebarContent
+            onClose={onClose}
+            display={{ base: "none", md: "block" }}
+            pt={20}
+          />
+          <Drawer
+            isOpen={isOpen}
+            placement="left"
+            onClose={onClose}
+            returnFocusOnClose={false}
+            onOverlayClick={onClose}
+            size="full"
+          >
+            <DrawerContent>
+              <SidebarContent onClose={onClose} />
+            </DrawerContent>
+          </Drawer>
+          <Box
+            pl={{ base: 2, md: 64 }}
+            pr={{ base: 2, md: 6 }}
+            pt={{ base: 2, md: 24 }}
+          >
             <Routes>
               <Route path="/" element={<Home />} />
               <Route path="/access-history" element={<AccessHistory />} />
               <Route path="/profile" element={<Profile />} />
             </Routes>
-          </BrowserRouter>
+          </Box>
         </Box>
-      </Box>
-      <Footer />
+        <Footer />
+      </BrowserRouter>
     </ChakraProvider>
   );
 };
