+++ frontend/src/routes/Profile/index.tsx
@@ -1,44 +1,54 @@
-import { Box, Grid } from '@chakra-ui/react';
-import { useEffect, useState } from 'react';
+import { Box, Grid } from "@chakra-ui/react";
+import { useEffect, useState } from "react";
+import { useLocation } from "react-router-dom";
 import { profileApi } from "../../api";
-import Banner from '../../features/Banner';
-import AvatarList from '../../features/AvatarList';
-import {
-    GetUserById200Response,
-    Avatar
-} from "../../schema";
+import Banner from "../../features/Banner";
+import AvatarList from "../../features/AvatarList";
+import { GetUserById200Response, Avatar } from "../../schema";
+
+type AuthUser = {
+  id: number;
+};
+
+const AUTH_USER: AuthUser = {
+  id: 9,
+};
 
 export default function Profile() {
   const [userData, setUserData] = useState<GetUserById200Response | null>(null);
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
+  const location = useLocation();
+  const userId = (location.state as { userId?: number })?.userId || AUTH_USER.id;
 
   useEffect(() => {
     async function fetchUserData() {
       try {
-        const response = await profileApi.getUserById(9);
+        const response = await profileApi.getUserById(userId);
         setUserData(response.data);
       } catch (err) {
-        setError('データの取得に失敗しました');
+        setError("データの取得に失敗しました");
       } finally {
         setLoading(false);
       }
     }
 
     fetchUserData();
-  }, []);
+  }, [userId]);
 
   const updateUserData = async (userId: number, avatarId: number) => {
     try {
+      if (userId === AUTH_USER.id) {
         const requestBody: Avatar = {
-            user_id: userId,
-            avatar_id: avatarId
-          };
-          await profileApi.putAvatar(requestBody);
-      const response = await profileApi.getUserById(9); // 再取得
-      setUserData(response.data);
+          user_id: userId,
+          avatar_id: avatarId,
+        };
+        await profileApi.putAvatar(requestBody);
+        const response = await profileApi.getUserById(userId); // 再取得
+        setUserData(response.data);
+      }
     } catch (err) {
-      setError('アバターの更新に失敗しました');
+      setError("アバターの更新に失敗しました");
     }
   };
 
@@ -48,20 +58,21 @@ export default function Profile() {
   if (!userData) return null;
 
   return (
-    <Box pt={{ base: '80px', md: '80px', xl: '10px' }}>
+    <Box pt={{ base: "80px", md: "80px", xl: "10px" }}>
       {/* Main Fields */}
       <Grid
         templateColumns={{
-          base: '1fr',
-          lg: '1.62fr 1fr'
+          base: "1fr",
+          lg: "1.62fr 1fr",
         }}
         templateRows={{
-          base: 'repeat(2, 1fr)',
-          lg: '1fr'
+          base: "repeat(2, 1fr)",
+          lg: "1fr",
         }}
-        gap={{ base: '20px', xl: '20px' }}>
+        gap={{ base: "20px", xl: "20px" }}
+      >
         <Banner
-          gridArea='1 / 1 / 2 / 2'
+          gridArea="1 / 1 / 2 / 2"
           banner="rgba(79, 209, 197, 1)"
           avatar={`./avatar/${userData.avatar_img_path}`}
           name={userData.user_name}
@@ -72,9 +83,11 @@ export default function Profile() {
           grade={userData.grade}
         />
         <AvatarList
-          gridArea={{ base: '2 / 1 / 3 / 2', lg: '1 / 2 / 2 / 3' }}
+          gridArea={{ base: "2 / 1 / 3 / 2", lg: "1 / 2 / 2 / 3" }}
           avatars={userData.avatar_list}
-          onAvatarClick={(avatarId) => updateUserData(userData.user_id, avatarId)}
+          onAvatarClick={(avatarId) =>
+            updateUserData(userData.user_id, avatarId)
+          }
         />
       </Grid>
     </Box>
